requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Data Quality & Validation
great-expectations>=0.17.0
pydantic>=2.0.0

# ML & Analytics
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Security
python-dotenv>=1.0.0
cryptography>=41.0.0
pyjwt>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
faker>=19.0.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pylint>=2.17.0

# Logging & Monitoring
python-json-logger>=2.0.7
sentry-sdk>=1.32.0